% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOFTest_betaSBM.R
\name{goftest}
\alias{goftest}
\title{Monte Carlo goodness-of-fit tests for Stochastic Blockmodels (SBMs)}
\usage{
goftest(A, K = NULL, C = NULL, numGraphs = 100)
}
\arguments{
\item{A}{a n by n binary symmetric adjacency matrix representing an undirected graph where n is the number of nodes in the graph}

\item{K}{a positive integer scalar representing the number of blocks; K>1}

\item{C}{a positive integer vector of size n for block assignments of each node; from 1 to K (no of blocks)}

\item{numGraphs}{number of graphs to be sampled; default value is 100}
}
\value{
A list with the elements
\item{statistic}{the values of the chi-square test statistics on each sampled graph}
\item{p.value}{the p-value for the test}
}
\description{
\code{goftest} performs chi square goodness-of-fit test for network data considering the model as beta SBM (Karwa et al. (2023))
}
\examples{
RNGkind(sample.kind = "Rounding")
set.seed(1729)

# We model a network with 3 even classes
n1 = 50
n2 = 50
n3 = 50

# Generating block assignments for each of the nodes
n = n1 + n2 + n3
class = rep(c(1, 2, 3), c(n1, n2, n3))

# Generating the adjacency matrix of the network
# Generate the matrix of connection probabilities
cmat = matrix(
  c(
    30, 0.05, 0.05,
    0.05, 30, 0.05,
    0.05, 0.05, 30
  ),
  ncol = 3,
  byrow = TRUE
)
pmat = cmat / n

# Creating the n x n adjacency matrix
adj <- matrix(0, n, n)
for (i in 2:n) {
  for (j in 1:(i - 1)) {
    p = pmat[class[i], class[j]] # We find the probability of connection with the weights
    adj[i, j] = rbinom(1, 1, p) # We include the edge with probability p
  }
}

adjsymm = adj + t(adj)

# When class assignment is known
out = goftest(adjsymm, C = class, numGraphs = 100)

chi_sq_seq = out$statistic
pvalue = out$p.value
print(pvalue)

# Plotting histogram of the sequence of the test statistics
hist(chi_sq_seq, 20, xlab = "chi-square test statistics", main = NULL)
abline(v = chi_sq_seq[1], col = "red", lwd = 5) # adding test statistic on the observed network

}
\references{
Karwa et al. (2023). "Monte Carlo goodness-of-fit tests for degree corrected and related stochastic blockmodels",
\emph{Journal of the Royal Statistical Society Series B: Statistical Methodology},
\url{https://doi.org/10.1093/jrsssb/qkad084}

Qin, T., and Rohe, K. (2013). "Regularized spectral clustering under the degree-corrected stochastic blockmodel",
\emph{Advances in neural information processing systems},
\url{https://proceedings.neurips.cc/paper_files/paper/2013/file/0ed9422357395a0d4879191c66f4faa2-Paper.pdf}

Lei, J., & Rinaldo, A. (2015). "Consistency of spectral clustering in stochastic block models",
\emph{The Annals of Statistics},
\url{https://doi.org/10.1214/14-AOS1274}

Li T, Levina E, & Zhu J (2021). "randnet: Random Network Model Estimation, Selection and Parameter Tuning".
\emph{R packageversion 0.3},
\url{https://CRAN.R-project.org/package=randnet}

Ghosh, S (2022). “MCGoSBM: Monte Carlo Goodness of Fit Tests for Stochastic Block Models”.
\emph{Github Repository}
\url{https://github.com/GhoshSoham/MCGoFSBM/tree/main}
}
